Update en-GB American word substitution
LanguageTool suggests the word "run" should be replaced by "ladder". As a native en-GB speaker, this makes no sense at all. The only thing I can think of that may have caused this confusion is that the step/bar of a ladder is called a "rung".

This removes false positives on phrases like "John likes to go running" and "I want to run the application".