[de] Fix bug #258 (Unnecessary whitespace triggers DE_CASE)
The reason for bug #258 (Unnecessary whitespace triggers DE_CASE): 
The first word in the sentence "   Im Folgenden ..." was not correctly 
tagged as it was not converted to lower case. (However, this was a 
consequence of the fact that a single white space was recognized as the 
first word and not "Im".)

Fix: GermanTagger now ignores all words consisting of white space 
characters only, when it determines the first word of a sentence.