cleanup enable/disable Rule/Category API (issue #421)
There are multiple issues with the enable/disable Rule/Category code, e.g.

 - There are some corner cases where `getAllActiveRules` and `ignoreRule`
   disagree about active rules.

 - Having two collections for disabled categories (by name and by id)
   makes the code more error-prone, e.g. the `getAllActiveRules`
   ignores `disabledCategories`.

 - Having `enableDefaultOffRule` and `enableRule` makes the API more
   cumbersome and in most cases we call both even if the rule is not
   default off.

LanguageTool uses `ignoreRule` internally to decide if a rule is active
or not. Refactor `ignoreRule` to use the logic of `getAllActiveRules`.
Refactor `getAllActiveRules` to use `ignoreRule`.

Remove `disabledCategories` field and refactor `disableCategory(String)`
to use `disableCategory(CategoryId)`.

Deprecate `enableDefaultOffRule` and `enableDefaultOffRuleCategory`,
they have the same functionality with enable Rule/Category now.

Internally, for rules, we had two collections (enabledRules,
disabledRules). `enabledRules` should contain default=off rules that had
explicitly enabled and `disabledRules` contained rules that were
disabled by the user. A rule could exists in both collections.

Now, a rule can only be in one collection. Explicitly enabled rules are
in `enabledRules`, and explicitly disabled rules are in `disabledRules`.
If a rule is not in one of them, then the rule will be active if not off
by default. The same logic goes for categories.