[en] Improve short message 'Grammatical problem'
Partly fixes
http://forum.languagetool.org/t/grammatical-problem-is-too-general-a-message/1235

I did these:
* Improved the short messages for most the rules that have the short
description 'Grammatical problem'. If I could easily give a short
instruction, I did that. Otherwise, I added a short description of the
problem. I kept 'Grammatical problem' in the description because the
rules are from the category 'GRAMMAR'.
* Added a marker to the correct examples so that a user can see a green
highlight in the example.
* Added some correct examples so that the correct and the incorrect
examples are the same, except for the error. The user can easily see the
problem and solution.

Some of the remaining rules with the short description 'Grammatical
problem' should be in the category NONSTANDARD_PHRASES. Most others are
an agreement error, and I intend to standardize the description for
those errors.