[gui] add line numbers to the text area of the stand-alone application
Credits: The implementation of this class is based on "Text Component
Line Number" Posted by Rob Camick on May 23, 2009 at
https://tips4java.wordpress.com/2009/05/23/text-component-line-number/

Usage / License according to https://tips4java.wordpress.com/about/ :
You are free to use and/or modify and/or distribute any or all code
posted on the Java Tips Weblog without restriction. A credit in the code
comments would be nice, but not in any way mandatory.