[en] Add coherency, POS, spelling (21)
Verified with www.collinsdictionary.com/spellcheck/english,
https://en.oxforddictionaries.com, www.merriam-webster.com.
For the nouns, if only one of the ~i[s|z]ation pair had a postag, I
added the same postag to the 'null' word. Usually, I did not check the
plural spellings. If neither noun had a postag and I was not sure about
the countability, I added only NN:U.