[en] add words to dictionary #1543 (small changes)
I removed or changed these in added.txt:
APIs: The base form is API, not APIs.
ASINs
OSX: The correct term is Mac OS X (https://en.wikipedia.org/wiki/MacOS). As best I know, 'OSX' (or 'OS X') alone is not a proper noun.
SSD: it is an abbreviation, not a proper noun. We don't have a postag for abbreviation. I changed to NN.

spelling.txt:
removed endeavour. It is not AmE (https://www.merriam-webster.com/dictionary/endeavour). It is already in BrE.