update classes to new object models
@jimregan
You can look at this commit to see the changes needed to update your
pull request.
Please, check is the message in LogainmRule still makes sense.
Also, even though I have not changed the tagger behaviour, it is splitting
words on underscores, as shown by the corrected test. Looks good to me,
but verify if that is the behaviour you intend.