[en] Add POS, spelling, coherency for 'onscreen'
www.merriam-webster.com/dictionary/on-screen
www.ldoceonline.com/dictionary/on-screen

I added a coherence pair because although it is possible for one form to be an adverb and the other form to be an adjective, it is unlikely. Most style guides are likely to have the same form for both the adverb and the adjective.