en ALL_OF_THE: soften the message, make the shortest example less controversial, make the rule work at the sentence start
We've got a complaint that this warning in a JetBrains IDE seems too harsch for an optional stylistic issue.
They also note that the example correction "she is going to all the parks" doesn't sound convincingly nicer.
So we propose to make it more explicit that the goal of this warning is conciseness and to add a shorter example (which would then be used by the IDE) which seems to sound nicer.